postgresql/requirements.txt
#Flask==3.1.2
#SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
#python-dotenv==1.2.1
#Werkzeug==3.1.5